+++ examples/ecommerce/components/DatalayerProvider.tsx
@@ -8,7 +8,7 @@ import {
   createClient,
   persistedExchange,
 } from '@/fuse/client'
-import React from 'react'
+import React, { Suspense } from 'react'
 
 export const DatalayerProvider = (props: any) => {
   const [client, ssr] = React.useMemo(() => {
@@ -26,7 +26,7 @@ export const DatalayerProvider = (props: any) => {
 
   return (
     <Provider client={client} ssr={ssr}>
-      {props.children}
+      <Suspense>{props.children}</Suspense>
     </Provider>
   )
 }