+++ examples/ecommerce/fuse/server.ts
@@ -1,9 +1,3 @@
 // This is a generated file!
 
-const files = require.context('../types', true, /.ts$/)
-files
-  .keys()
-  .filter((path: string) => path.includes('types/'))
-  .forEach(files)
-
 export * from 'fuse/next/server'