+++ examples/ecommerce/next.config.js
@@ -1,6 +1,6 @@
 const { nextFusePlugin } = require('fuse/next/plugin')
 
 /** @type {import('next').NextConfig} */
-const nextConfig = nextFusePlugin()({})
+const nextConfig = nextFusePlugin({})({})
 
 module.exports = nextConfig