+++ examples/spacex/fuse/graphql.ts
@@ -1,6 +1,4 @@
 /* eslint-disable */
-// This is a generated file!
-
 import { TypedDocumentNode as DocumentNode } from '@graphql-typed-document-node/core'
 export type Maybe<T> = T | null
 export type InputMaybe<T> = Maybe<T>