+++ website/src/components/Card.tsx
@@ -1,16 +1,24 @@
 import { cn } from '@/utils/tailwind'
 import { ComponentProps } from 'react'
 
-type CardProps = {} & ComponentProps<'div'>
+type CardProps = {
+  height?: 'full' | 'auto'
+} & ComponentProps<'div'>
 
-export function Card({ className, ...props }: CardProps) {
+export function Card({ className, height = 'full', ...props }: CardProps) {
   return (
-    <div className="relative h-full shrink-0 rounded-[16px] border border-transparent p-1 shadow-card ">
+    <div
+      className={cn(
+        'relative shrink-0 rounded-[16px] border border-transparent p-1 shadow-card',
+        `h-${height}`,
+      )}
+    >
       <div className="absolute inset-0 rounded-[16px] bg-dark-card-gradient" />
       <div className="absolute inset-[1px] rounded-[16px] bg-gravel-950" />
       <div
         className={cn(
-          'relative h-full rounded-[12px] border border-gravel-800 bg-gravel-900 px-[24px] py-[20px]',
+          'relative rounded-[12px] border border-gravel-800 bg-gravel-900 px-[24px] py-[20px]',
+          `h-${height}`,
           className,
         )}
         {...props}