+++ website/tailwind.config.ts
@@ -55,6 +55,8 @@ const config: Config = {
         'dark-card-gradient': `linear-gradient(180deg, rgba(218, 218, 218, 0.2) 0%,  rgba(218, 218, 218, 0) 90%)`,
         'steps-line-gradient':
           'linear-gradient(180deg, rgba(219, 254, 1, 0.00) -0.85%, #77DA05 18.91%, rgba(119, 218, 5, 0.00) 101.69%)',
+        'vertical-line-gradient':
+          'linear-gradient(180deg, rgba(242, 242, 242, 0) 0%, rgba(242, 242, 242, 0) 15%, rgba(242, 242, 242, 0.06) 100%)',
       },
       fontFamily: {
         sans: ['var(--font-geist-sans)'],