+++ website/theme.config.tsx
@@ -27,8 +27,9 @@ const themeConfig: DocsThemeConfig = {
   banner: {
     key: 'initial-release',
     text: (
-      <a href="http://stellate.co/blog/introducing-fuse-js" target="_blank">
-        🎉 Announcing Fuse.js: the open source data layer for Next.js →
+      <a href="http://stellate.co/blog/announcing-fuse-js" target="_blank">
+        🎉 Announcing Fuse.js: the opinionated framework for easily creating
+        typesafe data layers →
       </a>
     ),
   },